Me voooy que lastima pero adioooosss

+++ Academia.Proyecto.API/Academia.Proyecto.API/_Features/SucursalesXEmpleados/Dtos/SucursalesXEmpleadoDto.cs
@@ -14,7 +14,7 @@
 
         public string? SucursalNombre { get; set; }
 
-        public decimal? Kilometros { get; set; }
+        public int? Kilometros { get; set; }
 
         public int? UsuarioCreacionId { get; set; }
